Add $template part shortcode argument.
Added a shortcode argument as $template, which is the name of an additional template part - located in yourtheme/simple-user-listing/slug-author.php. The name in $template must match the "slug" part of the filename. If no template is specified, it defaults to content-author.php.

Example: [userlist template="members"] with a corresponding file of yourtheme/simple-user-listing/members-author.php