Revised adding $template part shortcode argument.
Added a shortcode argument as $template, which is the name of an additional template part - located in yourtheme/simple-user-listing/content-slug.php. The name in $template must match the "slug" part of the filename. If no template is specified, it defaults to content-author.php.

Example: [userlist template="member"] with a corresponding file of yourtheme/simple-user-listing/content-member.php